refactor(hydro_deploy): moved shared methods to trait
