+++ hydro_deploy/hydro_cli/src/lib.rs
@@ -22,6 +22,7 @@ use crate::core::hydroflow_crate::ports::HydroflowSource;
 
 mod cli;
 use hydro_deploy as core;
+use hydro_deploy::ssh::LaunchedSSHHost;
 
 static TOKIO_RUNTIME: std::sync::RwLock<Option<tokio::runtime::Runtime>> =
     std::sync::RwLock::new(None);
